requirements.txt
tensorflow
keras
pyyaml
